refactor: Move sdk.Dec to math package
