+++ math/dec_internal_test.go
@@ -1,6 +1,7 @@
-package types
+package math
 
 import (
+	"encoding/json"
 	"math/big"
 	"testing"
 
@@ -23,20 +24,20 @@ func (s *decimalInternalTestSuite) TestPrecisionMultiplier() {
 
 func (s *decimalInternalTestSuite) TestZeroDeserializationJSON() {
 	d := Dec{new(big.Int)}
-	err := cdc.UnmarshalJSON([]byte(`"0"`), &d)
+	err := json.Unmarshal([]byte(`"0"`), &d)
 	s.Require().Nil(err)
-	err = cdc.UnmarshalJSON([]byte(`"{}"`), &d)
+	err = json.Unmarshal([]byte(`"{}"`), &d)
 	s.Require().NotNil(err)
 }
 
 func (s *decimalInternalTestSuite) TestSerializationGocodecJSON() {
 	d := MustNewDecFromStr("0.333")
 
-	bz, err := cdc.MarshalJSON(d)
+	bz, err := json.Marshal(d)
 	s.Require().NoError(err)
 
 	d2 := Dec{new(big.Int)}
-	err = cdc.UnmarshalJSON(bz, &d2)
+	err = json.Unmarshal(bz, &d2)
 	s.Require().NoError(err)
 	s.Require().True(d.Equal(d2), "original: %v, unmarshalled: %v", d, d2)
 }
